+++ Casa.h
@@ -7,7 +7,7 @@ class Casa: public Locuinta{
 protected:
     std::string localitate{};
 
-    const float taxa_teren = 0.15; // se adauga la pretul standard al casei o taxa de 15% si apoi se trunchiaza
+    const double taxa_teren = 0.15; // se adauga la pretul standard al casei o taxa de 15% si apoi se trunchiaza
 
 public:
     // constructor fara parametri
