+++ main.cpp
@@ -3,9 +3,9 @@
 
 int Agentie::pret_mp_oras = 1000;
 int Agentie::pret_mp_tara = 800;
-float Agentie::valoare_distanta_mare = 0.8;
-float Agentie::valoare_distanta_medie = 1.05;
-float Agentie::valoare_distanta_mica = 1.3;
+double Agentie::valoare_distanta_mare = 0.8;
+double Agentie::valoare_distanta_medie = 1.05;
+double Agentie::valoare_distanta_mica = 1.3;
 
 int main() {
